- Worked on hlsl code for the terrain blending (now have proper blending if using the 'Core.dll' from the bin folder) - Refactored the code a bit and moved stuff to seperate classes - Updated the map rendering - Added more area locations - Experimented with trees and lights

+++ RPF6/Rpf6File.cs
@@ -1435,11 +1435,9 @@ namespace CodeX.Games.RDR1.RPF6
                     RSC85_TotalVSize = virtOrSize;
                     RSC85_TotalPSize = phys;
                 }
-                else
-                    RSC85_SetMemSizes(virtOrSize, phys);
+                else RSC85_SetMemSizes(virtOrSize, phys);
             }
-            else
-                Flag1 = (int)((long)Flag1 & 1073741824L | (long)virtOrSize & 3221225471L);
+            else Flag1 = (int)(Flag1 & 1073741824L | virtOrSize & 3221225471L);
         }
 
         public int GetCompactSize(int size)
